Add 'code' directive, to simplify (db . pack)
This both simplifies a common pattern *and* fixes a bug, since it uses the length of
the list parameter to update the location rather than the length of the output bytestring.
The two lengths should always be identical, however using the latter can sometimes result
in an infinite loop when referring to labels in code.