Add support for statics
Statics are blocks of ASM which are placed at the end, after your main block,
and which are defined and accessed using a string identifier.  They allow you
to break the program apart into multiple Haskell functions without having to
link those functions in manually.  You do have to be a bit careful with them
though.